Added a fixture, deleted the extra code, fixed the wait
