+++ pages/order_feed_page.py
@@ -1,5 +1,6 @@
 import allure
 import selenium
+from data import Data
 from pages.base_page import BasePage
 from selenium.webdriver.support import expected_conditions
 from selenium.webdriver.support.wait import WebDriverWait
@@ -9,7 +10,7 @@ class OrderFeedPage(BasePage):
 
     @allure.step('Поиск элемента из множества')
     def find_element_among_set_with_wait(self, locator, order):
-        WebDriverWait(self.driver, 20).until(
+        WebDriverWait(self.driver, Data.TIME_WAIT).until(
             expected_conditions.visibility_of_element_located(
             locator))
         orders = self.driver.find_elements(*locator)
